pyproject.toml
[project]
name = "b4b"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.121.2",
    "httpx>=0.28.1",
    "ipaddress>=1.0.23",
    "javascript>=1!1.2.6",
    "minify-html>=0.18.1",
    "pillow>=12.0.0",
    "pycryptodome>=3.23.0",
    "pyhaproxy>=0.3.7",
    "redis>=7.1.0",
    "requests>=2.32.5",
    "uvicorn[standard]>=0.38.0",
]
